refactor: decompose export.py into export/ subpackage (Phase 3)
Extract 1,905-LOC monolith into 5 focused modules:
- pdf_inventory.py: PDF inventory panel
- thumbnails.py: thumbnail cards, panels, poller
- pages.py: pages summary and subtab
- tab_script.py: export tab client-side JavaScript
- tab_assembly.py: main export tab assembly

All existing imports resolve via export/__init__.py re-exports.
