+++ src/studio_ui/components/studio/export/__init__.py
@@ -1,0 +1,26 @@
+"""Studio Export tab components (subpackage)."""
+
+from __future__ import annotations
+
+from .pages import _render_export_pages_subtab, render_export_pages_summary
+from .pdf_inventory import render_pdf_inventory_panel
+from .tab_assembly import render_studio_export_tab
+from .thumbnails import (
+    _thumbnail_card_id,
+    render_export_thumbnail_card,
+    render_export_thumbnails_loading_shell,
+    render_export_thumbnails_panel,
+    render_export_thumbs_poller,
+)
+
+__all__ = [
+    "_render_export_pages_subtab",
+    "_thumbnail_card_id",
+    "render_export_pages_summary",
+    "render_export_thumbnail_card",
+    "render_export_thumbnails_loading_shell",
+    "render_export_thumbnails_panel",
+    "render_export_thumbs_poller",
+    "render_pdf_inventory_panel",
+    "render_studio_export_tab",
+]
